+++ apps/backoffice-v2/src/domains/entities/hooks/useSelectNextCase/useSelectNextCase.tsx
@@ -1,0 +1,8 @@
+import { useCallback } from 'react';
+
+export const useSelectNextCase = () => {
+  return useCallback(() => {
+    // @TODO: Implement
+    throw new Error('Not implemented');
+  }, []);
+};