+++ apps/backoffice-v2/src/domains/entities/hooks/useSelectNextEntity/useSelectNextEntity.tsx
@@ -1,8 +1,0 @@
-import { useCallback } from 'react';
-
-export const useSelectNextEntity = () => {
-  return useCallback(() => {
-    // @TODO: Implement
-    throw new Error('Not implemented');
-  }, []);
-};