+++ apps/backoffice-v2/src/lib/blocks/hooks/useDocumentBlocks/useDocumentBlocks.tsx
@@ -12,7 +12,6 @@ import { FunctionComponent, useCallback, useMemo } from 'react';
 import { selectWorkflowDocuments } from '@/pages/Entity/selectors/selectWorkflowDocuments';
 import { useStorageFilesQuery } from '@/domains/storage/hooks/queries/useStorageFilesQuery/useStorageFilesQuery';
 import { useApproveTaskByIdMutation } from '@/domains/entities/hooks/mutations/useApproveTaskByIdMutation/useApproveTaskByIdMutation';
-import { getPostRemoveDecisionEventName } from '@/pages/Entity/get-post-remove-decision-event-name';
 import { useRemoveDecisionTaskByIdMutation } from '@/domains/entities/hooks/mutations/useRemoveDecisionTaskByIdMutation/useRemoveDecisionTaskByIdMutation';
 import { CommonWorkflowStates, StateTag } from '@ballerine/common';
 import { X } from 'lucide-react';
@@ -20,7 +19,6 @@ import { valueOrNA } from '@/common/utils/value-or-na/value-or-na';
 import { createBlocksTyped } from '@/lib/blocks/create-blocks-typed/create-blocks-typed';
 import { ctw } from '@/common/utils/ctw/ctw';
 import { getDocumentsSchemas } from '@/pages/Entity/utils/get-documents-schemas/get-documents-schemas';
-import { getPostDecisionEventName } from '@/lib/blocks/components/CallToActionLegacy/hooks/useCallToActionLegacyLogic/useCallToActionLegacyLogic';
 import { useDocumentPageImages } from '@/lib/blocks/hooks/useDocumentPageImages';
 import { motionBadgeProps } from '@/lib/blocks/motion-badge-props';
 import { useRejectTaskByIdMutation } from '@/domains/entities/hooks/mutations/useRejectTaskByIdMutation/useRejectTaskByIdMutation';
@@ -70,7 +68,6 @@ export const useDocumentBlocks = ({
 }) => {
   const issuerCountryCode = extractCountryCodeFromWorkflow(workflow);
   const documentsSchemas = getDocumentsSchemas(issuerCountryCode, workflow);
-  const postDecisionEventName = getPostDecisionEventName(workflow);
   const documents = useMemo(() => selectWorkflowDocuments(workflow), [workflow]);
   const documentPages = useMemo(
     () => documents.flatMap(({ pages }) => pages?.map(({ ballerineFileId }) => ballerineFileId)),
@@ -80,11 +77,8 @@ export const useDocumentBlocks = ({
   const documentPagesResults = useDocumentPageImages(documents, storageFilesQueryResult);
 
   const { mutate: mutateApproveTaskById, isLoading: isLoadingApproveTaskById } =
-    useApproveTaskByIdMutation(workflow?.id, postDecisionEventName);
-  const { isLoading: isLoadingRejectTaskById } = useRejectTaskByIdMutation(
-    workflow?.id,
-    postDecisionEventName,
-  );
+    useApproveTaskByIdMutation(workflow?.id);
+  const { isLoading: isLoadingRejectTaskById } = useRejectTaskByIdMutation(workflow?.id);
   const onMutateApproveTaskById = useCallback(
     ({
         taskId,
@@ -97,11 +91,7 @@ export const useDocumentBlocks = ({
         mutateApproveTaskById({ documentId: taskId, contextUpdateMethod }),
     [mutateApproveTaskById],
   );
-  const postRemoveDecisionEventName = getPostRemoveDecisionEventName(workflow);
-  const { mutate: onMutateRemoveDecisionById } = useRemoveDecisionTaskByIdMutation(
-    workflow?.id,
-    postRemoveDecisionEventName,
-  );
+  const { mutate: onMutateRemoveDecisionById } = useRemoveDecisionTaskByIdMutation(workflow?.id);
 
   return (
     documents?.flatMap(