+++ packages/common/src/index.ts
@@ -20,6 +20,7 @@ export {
   zodErrorToReadable,
   isNonEmptyArray,
   isType,
+  zodBuilder,
 } from './utils';
 
 export type { IErrorWithMessage } from './utils';