+++ packages/common/src/utils/index.ts
@@ -19,4 +19,5 @@ export { replaceNullsWithUndefined } from './replace-null-with-undefined';
 export { dump } from './dump';
 export { isNonEmptyArray } from './is-non-empty-array';
 export { isType } from './is-type';
+export { zodBuilder } from './zod-builder';
 export { type IErrorWithMessage } from './is-error-with-message';