+++ services/workflows-service/src/collection-flow/controllers/collection-flow.controller.ts
@@ -16,10 +16,10 @@ import { FinishFlowDto } from '@/collection-flow/dto/finish-flow.dto';
 import { GetFlowConfigurationInputDto } from '@/collection-flow/dto/get-flow-configuration-input.dto';
 import { UpdateContextInputDto } from '@/collection-flow/dto/update-context-input.dto';
 import { ApiExcludeController } from '@nestjs/swagger';
+import { BUILT_IN_EVENT, ARRAY_MERGE_OPTION } from '@ballerine/workflow-core';
 
 @Public()
 @UseTokenAuthGuard()
-@ApiExcludeController()
 @ApiExcludeController()
 @common.Controller('collection-flow')
 export class ColectionFlowController {
@@ -49,6 +49,7 @@ export class ColectionFlowController {
 
     try {
       const adapter = this.adapterManager.getAdapter(activeWorkflow.workflowDefinitionId);
+
       return {
         result: activeWorkflow ? adapter.serialize(activeWorkflow) : null,
       };
@@ -58,6 +59,7 @@ export class ColectionFlowController {
           `${activeWorkflow.workflowDefinitionId as string} is not supported.`,
         );
       }
+
       throw error;
     }
   }
@@ -107,11 +109,6 @@ export class ColectionFlowController {
     );
   }
 
-  @common.Put('')
-  async updateFlow(@common.Body() payload: UpdateFlowDto, @TokenScope() tokenScope: ITokenScope) {
-    return await this.service.updateWorkflowRuntimeData(payload, tokenScope);
-  }
-
   @common.Put('/language')
   async updateFlowLanguage(
     @common.Body() { language }: UpdateFlowLanguageDto,
@@ -130,9 +127,18 @@ export class ColectionFlowController {
     @common.Body() { context }: UpdateContextInputDto,
     @TokenScope() tokenScope: ITokenScope,
   ) {
-    return await this.workflowService.updateContextById(tokenScope.workflowRuntimeDataId, context, [
+    return await this.workflowService.event(
+      {
+        id: tokenScope.workflowRuntimeDataId,
+        name: BUILT_IN_EVENT.DEEP_MERGE_CONTEXT,
+        payload: {
+          newContext: context,
+          arrayMergeOption: ARRAY_MERGE_OPTION.BY_ID,
+        },
+      },
+      [tokenScope.projectId],
       tokenScope.projectId,
-    ]);
+    );
   }
 
   @common.Post('/send-event')
@@ -149,8 +155,8 @@ export class ColectionFlowController {
 
   @common.Post('resubmit')
   async resubmitFlow(@TokenScope() tokenScope: ITokenScope) {
-    return this.service.resubmitFlow(
-      tokenScope.workflowRuntimeDataId,
+    await this.workflowService.event(
+      { id: tokenScope.workflowRuntimeDataId, name: 'RESUBMITTED' },
       [tokenScope.projectId],
       tokenScope.projectId,
     );