+++ services/workflows-service/src/providers/file/file-provider/http-file.service.ts
@@ -31,11 +31,13 @@ export class HttpFileService implements IStreamableFileProvider {
 
     const fileBuffer = response.data;
     await fsPromises.writeFile(localFilePath, Buffer.from(fileBuffer));
+
     return localFilePath;
   }
 
   async isRemoteExists(remoteFileConfig: TRemoteUri): Promise<boolean> {
     const response = await this.client.head(remoteFileConfig);
+
     return response.status >= 200 && response.status < 300;
   }
 
@@ -50,6 +52,7 @@ export class HttpFileService implements IStreamableFileProvider {
     const response: AxiosResponse<Readable> = await axios.get(remoteUri, {
       responseType: 'stream',
     });
+
     return response.data;
   }