+++ services/workflows-service/src/workflow/workflow.controller.external.unit.test.ts
@@ -1,11 +1,10 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { CallHandler, ExecutionContext, HttpStatus, INestApplication } from '@nestjs/common';
 import request from 'supertest';
-import { ACGuard } from 'nest-access-control';
+// import { ACGuard } from 'nest-access-control';
 import { ACLModule } from '@/common/access-control/acl.module';
-import { AclFilterResponseInterceptor } from '@/common/access-control/interceptors/acl-filter-response.interceptor';
-import { AclValidateRequestInterceptor } from '@/common/access-control/interceptors/acl-validate-request.interceptor';
-
+// import { AclFilterResponseInterceptor } from '@/common/access-control/interceptors/acl-filter-response.interceptor';
+// import { AclValidateRequestInterceptor } from '@/common/access-control/interceptors/acl-validate-request.interceptor';
 import { WorkflowControllerExternal } from './workflow.controller.external';
 import { WorkflowService } from './workflow.service';
 import { EventEmitter2 } from '@nestjs/event-emitter';
@@ -15,6 +14,9 @@ import { EndUserService } from '@/end-user/end-user.service';
 import { WorkflowTokenService } from '@/auth/workflow-token/workflow-token.service';
 import { WorkflowDefinitionService } from '@/workflow-defintion/workflow-definition.service';
 import { AppLoggerService } from '@/common/app-logger/app-logger.service';
+import { PrismaService } from '@/prisma/prisma.service';
+import { WinstonLogger } from '@/common/utils/winston-logger/winston-logger';
+// import { AclFilterResponseInterceptor } from '@/common/access-control/interceptors/acl-filter-response.interceptor';
 
 const acGuard = {
   canActivate: () => {
@@ -75,16 +77,24 @@ describe('Workflow (external)', () => {
           provide: AppLoggerService,
           useValue: {} as AppLoggerService,
         },
+        {
+          provide: 'LOGGER',
+          useClass: WinstonLogger,
+        },
+        {
+          provide: PrismaService,
+          useValue: {} as PrismaService,
+        },
       ],
       controllers: [WorkflowControllerExternal],
       imports: [ACLModule],
     })
-      .overrideGuard(ACGuard)
-      .useValue(acGuard)
-      .overrideInterceptor(AclFilterResponseInterceptor)
-      .useValue(aclFilterResponseInterceptor)
-      .overrideInterceptor(AclValidateRequestInterceptor)
-      .useValue(aclValidateRequestInterceptor)
+      // .overrideGuard(ACGuard)
+      // .useValue(acGuard)
+      // .overrideInterceptor(AclFilterResponseInterceptor)
+      // .useValue(aclFilterResponseInterceptor)
+      // .overrideInterceptor(AclValidateRequestInterceptor)
+      // .useValue(aclValidateRequestInterceptor)
       .compile();
 
     app = moduleRef.createNestApplication();