+++ packages/react-pdf-toolkit/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ballerine/react-pdf-toolkit
 
+## 1.0.15
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/config@1.0.7
+  - @ballerine/ui@0.3.30
+
 ## 1.0.14
 
 ### Patch Changes